requirements.txt
Flask
google-genai
Pillow
requests
opencage